requirements.txt
aiogram>=3.22
python-dotenv>=1.0
